requirements.txt
pygame>=2.0.0
pychord>=1.2.0
pyfluidsynth>=1.3.0